punch/requirement.txt
bcrypt==4.0.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
dnspython==2.2.1
dominate==2.7.0
email-validator==1.2.1
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Datepicker==0.14
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
#greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
pip-search==0.0.12
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.1
rich==12.5.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
